pyproject.toml
[project]
name = "flea"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.1",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "redis>=5.2.1",
    "requests>=2.32.3",
]
